Remove IoT hub service client package from this repo (#1665)
- Remove unit and e2e tests for this package
- Fix some warnings in e2e test code including deleting unused/unreferenced methods
- Update readme files to point to the new repo
- Remove e2e test environment variables that were only used in IoT hub service client tests (AAD credentials)

+++ iot-e2e-tests/common/src/test/java/tests/integration/com/microsoft/azure/sdk/iot/iothub/methods/DirectMethodsTests.java
@@ -180,7 +180,7 @@ public class DirectMethodsTests extends DirectMethodsCommon
 
         // Assert
         assertNotNull(buildExceptionMessage("method result was null", testInstance.identity.getClient()), result);
-        assertEquals(buildExceptionMessage("Expected SUCCESS but got " + result.getStatus(), testInstance.identity.getClient()), (long)METHOD_SUCCESS, (long)result.getStatus());
+        assertEquals(buildExceptionMessage("Expected SUCCESS but got " + result.getStatus(), testInstance.identity.getClient()), METHOD_SUCCESS, (long)result.getStatus());
         assertEquals(buildExceptionMessage("Expected " + METHOD_DELAY_IN_MILLISECONDS + ":succeed" + " But got " + result.getPayload(String.class),
                 testInstance.identity.getClient()), METHOD_DELAY_IN_MILLISECONDS + ":succeed", result.getPayload(String.class));
     }
